LPS-112056 Move inner class ResourceActionsBag to portal-kernel.
