+++ portal-kernel/src/com/liferay/portal/kernel/security/permission/ResourceActionsBag.java
@@ -1,0 +1,39 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.kernel.security.permission;
+
+import java.util.Set;
+
+import org.osgi.annotation.versioning.ProviderType;
+
+/**
+ * @author Hai Yu
+ */
+@ProviderType
+public interface ResourceActionsBag {
+
+	public Set<String> getGroupDefaultActions();
+
+	public Set<String> getGuestDefaultActions();
+
+	public Set<String> getGuestUnsupportedActions();
+
+	public Set<String> getLayoutManagerActions();
+
+	public Set<String> getOwnerDefaultActions();
+
+	public Set<String> getSupportsActions();
+
+}
